[Refactor] Refactor shared-data storage volume mgr
