+++ fe/fe-core/src/main/java/com/starrocks/storagevolume/StorageVolume.java
@@ -105,6 +105,10 @@ public class StorageVolume {
         return id;
     }
 
+    public String getName() {
+        return name;
+    }
+
     public void setEnabled(boolean enabled) {
         this.enabled = enabled;
     }
